[Feat] : Add property subpath in MIddleTransportInfo
Git : #36
