+++ presentation/home/views/lib/sources/middle_transport/middle_transport_info/middle_transport_info.dart
@@ -1,0 +1,23 @@
+part of '../../../eb_home.dart';
+
+final class MiddleTransportInfo extends StatelessWidget {
+  final EBSubPath subPath;
+  final Stream<RealTimeInfo> stream;
+
+  const MiddleTransportInfo({
+    super.key,
+    required this.subPath,
+    required this.stream,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        _LeftDispatchColumn(subPath: subPath),
+        _RightDisPatchColumn(stream: stream),
+      ],
+    );
+  }
+}
