+++ presentation/home/views/lib/sources/middle_transport/middle_transport_view.dart
@@ -15,7 +15,9 @@ final class _MiddleTransportView extends StatelessWidget {
             content = const MiddleTransportAddSchedule();
           case InfoMiddleTransportState():
             final stream = middleState.realTimeInfoStream;
+            final subPath = middleState.subPath;
             content = MiddleTransportInfo(
+              subPath: subPath,
               stream: stream,
             );
           case OverScheduleMiddleTransportState():
